Поправил замечания

+++ project/include/utils.h
@@ -1,8 +1,9 @@
 #ifndef PROJECT_INCLUDE_UTILS_H_
 #define PROJECT_INCLUDE_UTILS_H_
+
 #include <stddef.h>
 
-size_t timer_from(int from);
+size_t timer_from(size_t from);
 int custom_pow(int base, int power);
 
 #endif  // PROJECT_INCLUDE_UTILS_H_
